#25: move Engine to classes
